requirements.txt
Flask
mysql-connector
dotenv
